requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
requests>=2.28.0
numpy>=1.21.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
